refactor: スラッシュコマンドを/list, /check, /createに分割
/reserve list, /reserve check, /reserve createのサブコマンド方式から
/list, /check, /createの独立コマンド方式に変更
