+++ discord-bot/src/commands/types.ts
@@ -1,0 +1,28 @@
+export interface InteractionOption {
+    name: string;
+    type: number;
+    value?: string;
+    options?: InteractionOption[];
+}
+
+export interface InteractionData {
+    name: string;
+    options?: InteractionOption[];
+}
+
+export interface InteractionUser {
+    id: string;
+}
+
+export interface InteractionMember {
+    user: InteractionUser;
+}
+
+export interface Interaction {
+    type: number;
+    data: InteractionData;
+    member?: InteractionMember;
+    user?: InteractionUser;
+    token: string;
+    id: string;
+}
